requirements.txt
lxml==4.7.1
Pillow==8.4.0
twython==3.9.1
urllib3==1.26.6
